cast R_xlen_t to long explicitly

+++ src/assign.c
@@ -467,7 +467,8 @@ SEXP assign(SEXP dt, SEXP rows, SEXP cols, SEXP newcolnames, SEXP values)
       // Can growVector at this point easily enough, but it shouldn't happen in first place so leave it as
       // strong error message for now.
     else if (TRUELENGTH(names) != oldtncol)
-      error(_("Internal error: selfrefnames is ok but tl names [%d] != tl [%d]"), TRUELENGTH(names), oldtncol);  // # nocov
+      // Use (long) to cast R_xlen_t to a fixed type to robustly avoid -Wformat compiler warnings
+      error(_("Internal error: selfrefnames is ok but tl names [%ld] != tl [%ld]"), (long)TRUELENGTH(names), (long)oldtncol);  // # nocov
     SETLENGTH(dt, oldncol+LENGTH(newcolnames));
     SETLENGTH(names, oldncol+LENGTH(newcolnames));
     for (i=0; i<LENGTH(newcolnames); i++)
